Changed source directory; Changed source repositories; Fixed wrong march-mabi pair; Added second stage gcc

+++ build.sh
@@ -13,20 +13,31 @@ __OPT_TARGET_MULTILIB="rv32e-ilp32e-- rv32ec-ilp32e-- rv32em-ilp32e-- rv32emc-il
   rv32ema-ilp32e-- rv32emac-ilp32e-- rv32ea-ilp32e-- rv32eac-ilp32e--
   rv32i-ilp32-- rv32ic-ilp32-- rv32im-ilp32-- rv32imc-ilp32--
   rv32ima-ilp32-- rv32imac-ilp32-- rv32ia-ilp32-- rv32iac-ilp32--
-  rv32if-ilp32-- rv32icf-ilp32-- rv32im-ilp32-- rv32imc-ilp32--
+  rv32if-ilp32-- rv32ifc-ilp32-- rv32im-ilp32-- rv32imc-ilp32--
   rv32imaf-ilp32-- rv32imafc-ilp32-- rv32iaf-ilp32-- rv32iafc-ilp32--"
 
+__VERSION_BINUTILS=binutils-2_31_1
+__VERSION_GDB=gdb-8.2.1-release
+__VERSION_GCC=gcc-8_2_0-release
+__VERSION_NEWLIB=newlib-3.0.0
+__VERSION_UCLIBCPP=v0.2.4
+
 export PATH=$PATH:${__OPT_TARGET_PATH}/bin
+__ROOT_DIR=`pwd`
 
 # ----------------------------------------------------------------------------
 # installation
 # ----------------------------------------------------------------------------
 
-sudo apt install -y build-essential flex bison texinfo
-sudo apt install -y linux-headers-amd64
-sudo apt install -y libgmp-dev libgmp10
-sudo apt install -y libmpfr-dev libmpfr4
-sudo apt install -y libmpc-dev libmpc3
+if [ ! -f .installed-libs ]; then
+  sudo apt install -y build-essential flex bison texinfo
+  sudo apt install -y linux-headers-amd64
+  sudo apt install -y libgmp-dev libgmp10
+  sudo apt install -y libmpfr-dev libmpfr4
+  sudo apt install -y libmpc-dev libmpc3
+  sudo apt install -y zlib1g-dev zlib1g
+fi
+touch .installed-libs
 
 # ----------------------------------------------------------------------------
 # sources
@@ -41,21 +52,21 @@ function git_checkout() {
   || git clone --config core.autocrlf=input --depth=1 --branch=$branch $repository $directory
 }
 
-__ROOT_DIR=`pwd`
-git_checkout 'riscv-binutils' 'riscv-next' 'https://github.com/riscv/riscv-binutils-gdb.git'
-git_checkout 'riscv-gcc' 'riscv-next' 'https://github.com/riscv/riscv-gcc.git'
-git_checkout 'riscv-newlib' 'riscv-newlib-next' 'https://github.com/riscv/riscv-newlib.git'
-git_checkout 'riscv-uclibc++' 'master' 'git://git.busybox.net/uClibc++'
+git_checkout 'src-binutils' __VERSION_BINUTILS 'git://sourceware.org/git/binutils-gdb.git'
+git_checkout 'src-gdb' __VERSION_GDB 'git://sourceware.org/git/binutils-gdb.git'
+git_checkout 'src-gcc' __VERSION_GCC 'git://gcc.gnu.org/git/gcc.git'
+git_checkout 'src-newlib' __VERSION_NEWLIB 'git://sourceware.org/git/newlib-cygwin.git'
+git_checkout 'src-uclibc++' __VERSION_UCLIBCPP 'git://git.busybox.net/uClibc++'
 
 # ----------------------------------------------------------------------------
 # binutils
 # ----------------------------------------------------------------------------
 
 cd $__ROOT_DIR
-if [ ! -d built-binutils ]; then
+if [ ! -f .built-binutils ]; then
   rm -rf build-binutils || true
   mkdir build-binutils && cd build-binutils
-  ../riscv-binutils/configure \
+  ../src-binutils/configure \
     --target=${__OPT_TARGET_ARCH} \
     --prefix=${__OPT_TARGET_PATH} \
     --with-arch=${__OPT_TARGET_MARCH_FULL} \
@@ -66,24 +77,24 @@ if [ ! -d built-binutils ]; then
   make all ${__OPT_MULTICORE}
   make install ${__OPT_MULTICORE}
 fi
-mkdir $__ROOT_DIR/built-binutils | true
+touch $__ROOT_DIR/.built-binutils
 
 # ----------------------------------------------------------------------------
 # gcc
 # ----------------------------------------------------------------------------
 
 cd $__ROOT_DIR
-if [ ! -d built-gcc ]; then
-  if [ ! -f ./riscv-gcc/gcc/config/riscv/t-elf-multilib64 ]; then
-    mv ./riscv-gcc/gcc/config/riscv/t-elf-multilib ./riscv-gcc/gcc/config/riscv/t-elf-multilib64 | true
-    ./riscv-gcc/gcc/config/riscv/multilib-generator ${__OPT_TARGET_MULTILIB} \
-      > ./riscv-gcc/gcc/config/riscv/t-elf-multilib32
-    patch ./riscv-gcc/gcc/config.gcc ./config.gcc.patch
+if [ ! -f .built-gcc ]; then
+  if [ ! -f ./src-gcc/gcc/config/riscv/t-elf-multilib64 ]; then
+    mv ./src-gcc/gcc/config/riscv/t-elf-multilib ./src-gcc/gcc/config/riscv/t-elf-multilib64 | true
+    ./src-gcc/gcc/config/riscv/multilib-generator ${__OPT_TARGET_MULTILIB} \
+      > ./src-gcc/gcc/config/riscv/t-elf-multilib32
+    patch ./src-gcc/gcc/config.gcc ./config.gcc.patch
   fi
 
   rm -rf build-gcc || true
   mkdir build-gcc && cd build-gcc
-  ../riscv-gcc/configure \
+  ../src-gcc/configure \
     --target=${__OPT_TARGET_ARCH} \
     --prefix=${__OPT_TARGET_PATH} \
     --program-prefix=${__OPT_TARGET_PREFIX} \
@@ -92,45 +103,55 @@ if [ ! -d built-gcc ]; then
     --with-arch=${__OPT_TARGET_MARCH} --with-abi=${__OPT_TARGET_MABI} \
     --enable-lto --enable-multilib \
     --disable-nls --disable-wchar_t --disable-threads --disable-libstdcxx \
-    --enable-initfini-array
+    --enable-initfini-array \
+    --with-system-zlib --with-gnu-as --with-gnu-ld
   make all-gcc ${__OPT_MULTICORE}
   make install-gcc ${__OPT_MULTICORE}
 fi
-mkdir $__ROOT_DIR/built-gcc | true
+touch $__ROOT_DIR/.built-gcc
 
 # ----------------------------------------------------------------------------
 # checking gcc
 # ----------------------------------------------------------------------------
 
-cd $__ROOT_DIR
-if [ ! -d checked-gcc ]; then
-  rm -rf check-gcc || true
-  mkdir check-gcc && cd check-gcc
-
-  cat >start.s <<EOF
-.section .text
-_start:
-  nop
-EOF
-
-  for type in ${__OPT_TARGET_MULTILIB}; do
-    march=`echo $type | sed 's/\(.*\)-\(.*\)--/\1/'`
-    mabi=`echo $type | sed 's/\(.*\)-\(.*\)--/\2/'`
-
-    ${__OPT_TARGET_PREFIX}-gcc --march=$march --mabi=$mabi -o $march.o start.s
-  done
-fi
-mkdir $__ROOT_DIR/checked-gcc | true
+#cd $__ROOT_DIR
+#if [ ! -d checked-gcc ]; then
+#  rm -rf check-gcc || true
+#  mkdir check-gcc && cd check-gcc
+#
+#  __LDFLAGS='-nostdlib -Wl,-Bstatic,-T,../test/link.ld,--no-gc-sections'
+#  __CCFLAGS='-Os --std=c99 -ffreestanding -nostdlib'
+#
+#    ${__OPT_TARGET_PREFIX}gcc ${__CCFLAGS} -o start.o -c ../test/start.c
+#    ${__OPT_TARGET_PREFIX}gcc ${__CCFLAGS} -o start-asm.o -c ../test/start.s
+#    ${__OPT_TARGET_PREFIX}gcc ${__LDFLAGS} -o start.elf start.o
+#    ${__OPT_TARGET_PREFIX}gcc ${__LDFLAGS} -o start-asm.elf start-asm.o
+#    ${__OPT_TARGET_PREFIX}readelf start.elf -a > start.txt
+#    ${__OPT_TARGET_PREFIX}readelf start-asm.elf -a > start-asm.txt
+#
+#  for type in ${__OPT_TARGET_MULTILIB}; do
+#    march=`echo $type | sed 's/\(.*\)-\(.*\)--/\1/'`
+#    mabi=`echo $type | sed 's/\(.*\)-\(.*\)--/\2/'`
+#
+#    ${__OPT_TARGET_PREFIX}gcc ${__CCFLAGS} -march=$march -mabi=$mabi -o $march.o -c ../test/start.c
+#    ${__OPT_TARGET_PREFIX}gcc ${__CCFLAGS} -march=$march -mabi=$mabi -o $march-asm.o -c ../test/start.s
+#    ${__OPT_TARGET_PREFIX}gcc ${__LDFLAGS} -o $march.elf $march.o
+#    ${__OPT_TARGET_PREFIX}gcc ${__LDFLAGS} -o $march-asm.elf $march-asm.o
+#    ${__OPT_TARGET_PREFIX}readelf $march.elf -a > $march.txt
+#    ${__OPT_TARGET_PREFIX}readelf $march-asm.elf -a > $march-asm.txt
+#  done
+#fi
+#mkdir $__ROOT_DIR/checked-gcc | true
 
 # ----------------------------------------------------------------------------
 # newlib (libc)
 # ----------------------------------------------------------------------------
 
 cd $__ROOT_DIR
-if [ ! -d built-newlib ]; then
+if [ ! -f .built-newlib ]; then
   rm -rf build-newlib || true
   mkdir build-newlib && cd build-newlib
-  ../riscv-newlib/configure \
+  ../src-newlib/configure \
     --target=${__OPT_TARGET_ARCH} \
     --with-arch=${__OPT_TARGET_MARCH} --with-abi=${__OPT_TARGET_MABI} \
     --prefix=${__OPT_TARGET_PATH} \
@@ -141,14 +162,39 @@ if [ ! -d built-newlib ]; then
   make all ${__OPT_MULTICORE}
   make install ${__OPT_MULTICORE}
 fi
-mkdir $__ROOT_DIR/built-newlib | true
+touch $__ROOT_DIR/.built-newlib
+
+# ----------------------------------------------------------------------------
+# gcc stage 2
+# ----------------------------------------------------------------------------
+
+cd $__ROOT_DIR
+if [ ! -f .built-gcc-stage2 ]; then
+  rm -rf build-gcc-stage2 || true
+  mkdir build-gcc-stage2 && cd build-gcc-stage2
+  ../src-gcc/configure \
+    --target=${__OPT_TARGET_ARCH} \
+    --prefix=${__OPT_TARGET_PATH} \
+    --program-prefix=${__OPT_TARGET_PREFIX} \
+    ${__OPT_BOOTSTRAP} \
+    --without-headers --enable-languages=c,c++ \
+    --with-arch=${__OPT_TARGET_MARCH} --with-abi=${__OPT_TARGET_MABI} \
+    --enable-lto --enable-multilib \
+    --disable-nls --disable-wchar_t --disable-threads --disable-libstdcxx \
+    --enable-initfini-array \
+    --with-system-zlib --with-newlib --disable-shared \
+    --with-gnu-as --with-gnu-ld
+  make all ${__OPT_MULTICORE}
+  make install ${__OPT_MULTICORE}
+fi
+touch $__ROOT_DIR/.built-gcc-stage2
 
 # ----------------------------------------------------------------------------
 # uclibc++ (libcpp)
 # ----------------------------------------------------------------------------
 
 function build_uclibcpp() {
-  MAKE_PARAMS='CROSS_COMPILE="${__OPT_TARGET_PREFIX}-"'
+  MAKE_PARAMS='CROSS_COMPILE="${__OPT_TARGET_PREFIX}"'
   if [ -n "$1" && -n "$2" ]; then
     MAKE_PARAMS+=' HOSTCFLAGS="-march=$1 -mabi=$2"'
     MAKE_PARAMS+=' HOSTCXXFLAGS="-march=$1 -mabi=$2"'
@@ -170,8 +216,8 @@ function build_uclibcpp() {
 }
 
 cd $__ROOT_DIR
-if [ ! -d built-uclibc++ ]; then
-  cd riscv-uclibc++
+if [ ! -f .built-uclibc++ ]; then
+  cd src-uclibc++
   git reset --hard HEAD
   git clean -dfx
 
@@ -196,7 +242,7 @@ if [ ! -d built-uclibc++ ]; then
     mv $file $(dirname $file)/`echo $(basename $file) | tr [:upper:] [:lower:]`
   done;
 fi
-mkdir $__ROOT_DIR/built-uclibc++ | true
+touch $__ROOT_DIR/.built-uclibc++ | true
 
 cd $__ROOT_DIR
 echo "Done"
